+++ src/maybe.ts
@@ -12,35 +12,29 @@ import type { Monad, Comonad } from './types';
 class Maybe<V> implements Comonad, Monad {
   // TODO: review this when ECMAScript's private class fields will be
   // widely spread in browsers.
-  private readonly _value: V | null | undefined;
+  private readonly _value: V;
 
-  private constructor(value: V | null | undefined) {
+  private constructor(value: V) {
     this._value = value;
   }
 
-  /** Wraps value with `Maybe` monad with **Just** state. */
-  static just<T>(value: T): Maybe<T> {
-    return new Maybe<T>(value);
-  }
-
   /** Creates `Maybe` monad instance with **Nothing** state. */
-  static nothing<T = null>(): Maybe<T> {
-    return new Maybe<T>(null);
+  static nothing<T = unknown>(): Maybe<T> {
+    // @ts-ignore - container may have empty inner value, but
+    // we suppose that container can contain value of type `T`.
+    return Maybe.maybe(null);
   }
 
   /**
-   * Wraps value with `Maybe` monad. Function detects state
-   * (**Just** or **Nothing**) of `Maybe` by yourself.
+   * Wraps value with `Maybe` monad.
    * If value is `Maybe`, then its copy will be returned.
    */
-  static maybe<T>(value: T | Maybe<T> | null | undefined): Maybe<T> {
+  static maybe<T>(value: T | Maybe<T>): Maybe<T> {
     const exposedValue = isMaybe<T>(value) ? value.extract() : value;
-    return isNothing(exposedValue)
-      ? Maybe.nothing<T>()
-      : Maybe.just(exposedValue);
+    return new Maybe(exposedValue);
   }
 
-  isJust(): this is Maybe<V> {
+  isJust(): this is Maybe<NonNullable<V>> {
     return !this.isNothing();
   }
 
@@ -48,31 +42,25 @@ class Maybe<V> implements Comonad, Monad {
     return isNothing(this._value);
   }
 
-  map<R>(fn: (value: NonNullable<V>) => R | null | undefined): Maybe<R> {
-    return this.isJust()
-      ? Maybe.maybe(fn(this._value as NonNullable<V>))
-      : Maybe.nothing();
+  map<R>(fn: (value: NonNullable<V>) => R): Maybe<R> {
+    return this.isJust() ? Maybe.maybe(fn(this._value)) : Maybe.nothing();
   }
 
-  apply<R>(
-    other: Maybe<(value: NonNullable<V>) => R | null | undefined>
-  ): Maybe<R> {
-    return other.isJust()
-      ? this.map(other.extract() as (value: NonNullable<V>) => R)
-      : Maybe.nothing();
+  apply<R>(other: Maybe<(value: NonNullable<V>) => R>): Maybe<R> {
+    return other.isJust() ? this.map(other.extract()) : Maybe.nothing();
   }
 
   chain<R>(fn: (value: NonNullable<V>) => Maybe<R>): Maybe<R> {
-    return this.isJust() ? fn(this._value as NonNullable<V>) : Maybe.nothing();
+    return this.isJust() ? fn(this._value) : Maybe.nothing();
   }
 
-  extract(): V | null | undefined {
+  extract(): V {
     return this._value;
   }
 }
 
 export type { Maybe };
-export const { just, nothing, maybe } = Maybe;
+export const { maybe, nothing } = Maybe;
 
 /** Checks if value is instance of `Maybe` monad. */
 export function isMaybe<T>(value: any): value is Maybe<T> {